+++ srcs/utils_termios.c
@@ -6,7 +6,7 @@
 /*   By: acauchy <acauchy@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/04 14:19:50 by acauchy           #+#    #+#             */
-/*   Updated: 2018/06/11 14:53:46 by acauchy          ###   ########.fr       */
+/*   Updated: 2018/06/13 15:22:32 by acauchy          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void	disable_raw_mode(void)
 		exit_error("Can't set terminal back to default mode.");
 }
 
-size_t	get_term_cols(void)
+int		get_term_cols(void)
 {
 	struct ttysize	ts;
